+++ libs/github/go.mod
@@ -1,0 +1,15 @@
+module github.com/gravitational/shared-workflows/libs/github
+
+go 1.22.4
+
+require github.com/gravitational/trace v1.3.1
+
+require (
+	github.com/jonboulle/clockwork v0.4.0 // indirect
+	github.com/sirupsen/logrus v1.9.3 // indirect
+	github.com/stretchr/testify v1.8.4 // indirect
+	golang.org/x/crypto v0.17.0 // indirect
+	golang.org/x/net v0.17.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/term v0.15.0 // indirect
+)
